refactor(channel): move serve_runtime.rs to runtime/serve.rs
- Move serve_runtime.rs to channel/runtime/serve.rs
- Update internal imports: super::runtime_state -> super::super::runtime_state
- Update external imports in mod.rs, dispatch.rs, commands/serve.rs, whatsapp/mod.rs
- Change pub(super) to pub(in crate::channel) for visibility across runtime module boundary
