--- crates/app/src/channel/serve_runtime.rs
@@ -1,256 +1,0 @@
-use std::{
-    future::Future,
-    sync::{
-        Arc,
-        atomic::{AtomicBool, Ordering},
-    },
-    time::{SystemTime, UNIX_EPOCH},
-};
-
-use tokio::sync::Notify;
-
-use super::registry::ChannelCommandFamilyDescriptor;
-use super::runtime_state;
-use super::runtime_state::ChannelOperationRuntimeTracker;
-use super::types::ChannelPlatform;
-use crate::CliResult;
-
-#[cfg(any(
-    feature = "channel-telegram",
-    feature = "channel-feishu",
-    feature = "channel-matrix",
-    feature = "channel-wecom",
-    feature = "channel-whatsapp"
-))]
-#[derive(Debug, Clone, Copy)]
-pub(super) struct ChannelServeRuntimeSpec<'a> {
-    pub(super) platform: ChannelPlatform,
-    pub(super) operation_id: &'static str,
-    pub(super) account_id: &'a str,
-    pub(super) account_label: &'a str,
-}
-
-#[cfg(any(
-    feature = "channel-telegram",
-    feature = "channel-feishu",
-    feature = "channel-matrix",
-    feature = "channel-wecom",
-    feature = "channel-whatsapp"
-))]
-#[derive(Debug, Clone, Copy)]
-pub(super) struct ChannelServeCommandSpec {
-    pub(super) family: ChannelCommandFamilyDescriptor,
-}
-
-#[derive(Debug, Clone)]
-pub struct ChannelServeStopHandle {
-    requested: Arc<AtomicBool>,
-    stop: Arc<Notify>,
-}
-
-impl ChannelServeStopHandle {
-    pub fn new() -> Self {
-        Self {
-            requested: Arc::new(AtomicBool::new(false)),
-            stop: Arc::new(Notify::new()),
-        }
-    }
-
-    pub fn request_stop(&self) {
-        self.requested.store(true, Ordering::SeqCst);
-        self.stop.notify_waiters();
-    }
-
-    pub fn is_requested(&self) -> bool {
-        self.requested.load(Ordering::SeqCst)
-    }
-
-    #[cfg(any(
-        feature = "channel-telegram",
-        feature = "channel-feishu",
-        feature = "channel-matrix",
-        feature = "channel-wecom",
-        feature = "channel-whatsapp"
-    ))]
-    pub(super) async fn wait(&self) {
-        if self.is_requested() {
-            return;
-        }
-        let notified = self.stop.notified();
-        tokio::pin!(notified);
-        if self.is_requested() {
-            return;
-        }
-        notified.await;
-    }
-}
-
-#[cfg(any(
-    feature = "channel-telegram",
-    feature = "channel-feishu",
-    feature = "channel-matrix",
-    feature = "channel-wecom",
-    feature = "channel-whatsapp"
-))]
-pub(super) fn channel_runtime_now_ms() -> u64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|value| value.as_millis() as u64)
-        .unwrap_or(0)
-}
-
-#[cfg(any(
-    feature = "channel-telegram",
-    feature = "channel-feishu",
-    feature = "channel-matrix",
-    feature = "channel-wecom",
-    feature = "channel-whatsapp"
-))]
-pub(super) fn ensure_channel_operation_runtime_slot_available_in_dir(
-    runtime_dir: &std::path::Path,
-    spec: ChannelServeRuntimeSpec<'_>,
-) -> CliResult<()> {
-    let now = channel_runtime_now_ms();
-    runtime_state::prune_inactive_channel_operation_runtime_files_for_account_from_dir(
-        runtime_dir,
-        spec.platform,
-        spec.operation_id,
-        Some(spec.account_id),
-        now,
-    )?;
-    let Some(runtime) = runtime_state::load_channel_operation_runtime_for_account_from_dir(
-        runtime_dir,
-        spec.platform,
-        spec.operation_id,
-        spec.account_id,
-        now,
-    ) else {
-        return Ok(());
-    };
-    if runtime.running_instances == 0 {
-        return Ok(());
-    }
-
-    let pid = runtime
-        .pid
-        .map(|value| value.to_string())
-        .unwrap_or_else(|| "unknown".to_owned());
-    Err(format!(
-        "{} account `{}` already has an active {} runtime (pid={}, running_instances={}); stop the existing instance or wait for it to become stale before restarting",
-        spec.platform.as_str(),
-        spec.account_id,
-        spec.operation_id,
-        pid,
-        runtime.running_instances
-    ))
-}
-
-#[cfg(any(
-    feature = "channel-telegram",
-    feature = "channel-feishu",
-    feature = "channel-matrix",
-    feature = "channel-wecom",
-    feature = "channel-whatsapp"
-))]
-pub(super) async fn with_channel_serve_runtime<T, F, Fut>(
-    spec: ChannelServeRuntimeSpec<'_>,
-    run: F,
-) -> CliResult<T>
-where
-    F: FnOnce(Arc<ChannelOperationRuntimeTracker>) -> Fut,
-    Fut: Future<Output = CliResult<T>>,
-{
-    ensure_channel_operation_runtime_slot_available_in_dir(
-        runtime_state::default_channel_runtime_state_dir().as_path(),
-        spec,
-    )?;
-    let runtime = Arc::new(
-        ChannelOperationRuntimeTracker::start(
-            spec.platform,
-            spec.operation_id,
-            spec.account_id,
-            spec.account_label,
-        )
-        .await?,
-    );
-    let result = run(runtime.clone()).await;
-    let shutdown_result = runtime.shutdown().await;
-    match result {
-        Err(error) => Err(error),
-        Ok(value) => {
-            shutdown_result?;
-            Ok(value)
-        }
-    }
-}
-
-#[cfg(any(
-    feature = "channel-telegram",
-    feature = "channel-feishu",
-    feature = "channel-matrix",
-    feature = "channel-wecom",
-    feature = "channel-whatsapp"
-))]
-pub(super) async fn with_channel_serve_runtime_with_stop<F, Fut>(
-    spec: ChannelServeRuntimeSpec<'_>,
-    stop: ChannelServeStopHandle,
-    run: F,
-) -> CliResult<()>
-where
-    F: FnOnce(Arc<ChannelOperationRuntimeTracker>, ChannelServeStopHandle) -> Fut,
-    Fut: Future<Output = CliResult<()>>,
-{
-    with_channel_serve_runtime(spec, move |runtime| run(runtime, stop)).await
-}
-
-#[cfg(test)]
-pub(super) async fn with_channel_serve_runtime_with_stop_in_dir<F, Fut>(
-    runtime_dir: &std::path::Path,
-    process_id: u32,
-    spec: ChannelServeRuntimeSpec<'_>,
-    stop: ChannelServeStopHandle,
-    run: F,
-) -> CliResult<()>
-where
-    F: FnOnce(Arc<ChannelOperationRuntimeTracker>, ChannelServeStopHandle) -> Fut,
-    Fut: Future<Output = CliResult<()>>,
-{
-    with_channel_serve_runtime_in_dir(runtime_dir, process_id, spec, move |runtime| {
-        run(runtime, stop)
-    })
-    .await
-}
-
-#[cfg(test)]
-pub(super) async fn with_channel_serve_runtime_in_dir<T, F, Fut>(
-    runtime_dir: &std::path::Path,
-    process_id: u32,
-    spec: ChannelServeRuntimeSpec<'_>,
-    run: F,
-) -> CliResult<T>
-where
-    F: FnOnce(Arc<ChannelOperationRuntimeTracker>) -> Fut,
-    Fut: Future<Output = CliResult<T>>,
-{
-    ensure_channel_operation_runtime_slot_available_in_dir(runtime_dir, spec)?;
-    let runtime = Arc::new(
-        runtime_state::start_channel_operation_runtime_tracker_for_test(
-            runtime_dir,
-            spec.platform,
-            spec.operation_id,
-            spec.account_id,
-            spec.account_label,
-            process_id,
-        )
-        .await?,
-    );
-    let result = run(runtime.clone()).await;
-    let shutdown_result = runtime.shutdown().await;
-    match result {
-        Err(error) => Err(error),
-        Ok(value) => {
-            shutdown_result?;
-            Ok(value)
-        }
-    }
-}
